requirements.txt
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.0
decorator==5.1.1
executing==2.1.0
greenlet==3.1.1
idna==3.10
ipython==8.31.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
playwright==1.49.1
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pyee==12.0.0
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
